Requirement: add new Bookmarks
As a user
So that I can add new favourite websites
I want to add a bookmark to my bookmarks

Domain Model
Nouns: user, favourite websites, bookmark, list of bookmarks
Verbs: add

Created a user story based on a new requirement to add bookmarks into the database. Wrote a feature test and a matching unit test for the self.create method in the Bookmark class. Kept the controller skinny by connection to the database and adding a bookmark via the app.